requirements.txt.txt
selenium==4.21.0
requests==2.31.0
Pillow==10.0.0
tensorflow==2.13.0
numpy==1.23.0
opencv-python==4.6.0.66
